refactor insertion index to be publicly accessible
