+++ v2/internal/store/indexcheck.go
@@ -3,6 +3,7 @@ package store
 import (
 	"github.com/ipfs/go-cid"
 	carv2 "github.com/ipld/go-car/v2"
+	"github.com/ipld/go-car/v2/index"
 )
 
 // ShouldPut returns true if the block should be put into the CAR according to the options provided
@@ -10,7 +11,7 @@ import (
 // is an identity block and StoreIdentityCIDs is false, or because it already exists and
 // BlockstoreAllowDuplicatePuts is false.
 func ShouldPut(
-	idx *InsertionIndex,
+	idx *index.InsertionIndex,
 	c cid.Cid,
 	maxIndexCidSize uint64,
 	storeIdentityCIDs bool,
@@ -60,7 +61,7 @@ func ShouldPut(
 // rules associated with the options. Similar to ShouldPut, but for the simpler
 // Has() case.
 func Has(
-	idx *InsertionIndex,
+	idx *index.InsertionIndex,
 	c cid.Cid,
 	maxIndexCidSize uint64,
 	storeIdentityCIDs bool,
